+++ locale/en-us/script.lua
@@ -39,6 +39,7 @@ DIAG_UNBALANCED_ASSIGNMENTS = 'The value is assigned as `nil` because the number
 DIAG_REQUIRE_LIKE       = 'You can treat `{}` as `require` by setting.'
 DIAG_COSE_NON_OBJECT    = 'Cannot close a value of this type. (Unless set `__close` meta method)'
 DIAG_COUNT_DOWN_LOOP    = 'Do you mean `{}` ?'
+DIAG_IMPLICIT_ANY       = 'Can not infer type.'
 
 DIAG_CIRCLE_DOC_CLASS                 = 'Circularly inherited classes.'
 DIAG_DOC_FIELD_NO_CLASS               = 'The field must be defined after the class.'